+++ pyxem/signals/indexation_results.py
@@ -493,11 +493,14 @@ def vectors_from_orientation_map(
     rotation = Rotation.from_euler(
         (mirror * rotation, 0, 0), degrees=True, direction="crystal2lab"
     )
-
+    coordinate_format = vectors.coordinate_format
     vectors = ~rotation * vectors.to_miller()
     vectors = DiffractingVector(
-        vectors.phase, xyz=vectors.data.copy(), intensity=intensities
+        vectors.phase,
+        xyz=-vectors.data.copy(),  # Negating for proper alignment - is this flipping z direction?
+        intensity=intensities,
     )
+    vectors.coordinate_format = coordinate_format
     vectors.original_hkl = hkl
 
     # angle = np.deg2rad(mirror * rotation)
@@ -979,7 +982,7 @@ class OrientationMap(DiffractionVectors2D):
         annotation_shift: Sequence[float] = None,
         text_kwargs: dict = None,
         include_intensity: bool = False,
-        intesity_scale: float = 1,
+        intensity_scale: float = 1,
         fast: bool = True,
         **kwargs,
     ) -> Sequence[hs.plot.markers.Markers]:
@@ -1025,15 +1028,16 @@ class OrientationMap(DiffractionVectors2D):
         all_markers = []
         for n in range(n_best):
             vectors = self.to_vectors(
+                n_best_index=n,
+                return_object=False,
                 lazy_output=True,
                 navigation_chunks=navigation_chunks,
-                return_object=False,
             )
             color = marker_colors[n % len(marker_colors)]
             if include_intensity:
                 intensity = vectors.map(
                     vectors_to_intensity,
-                    scale=intesity_scale,
+                    scale=intensity_scale,
                     inplace=False,
                     ragged=True,
                     output_dtype=object,
@@ -1081,9 +1085,28 @@ class OrientationMap(DiffractionVectors2D):
             all_markers = compute_markers(all_markers)
         return all_markers
 
+    @deprecated(
+        since="0.20.0",
+        alternative="pyxem.signals.OrientationMap.to_polar_markers",
+        removal="1.0.0",
+    )
     def to_single_phase_polar_markers(
         self,
         signal_axes: Sequence[BaseDataAxis],
+        n_best: int = 1,
+        marker_colors: str = ("red", "blue", "green", "orange", "purple"),
+        lazy_output: bool = None,
+        **kwargs,
+    ) -> Iterator[hs.plot.markers.Markers]:
+        return self.to_polar_markers(
+            n_best=n_best,
+            marker_colors=marker_colors,
+            lazy_output=lazy_output,
+            **kwargs,
+        )
+
+    def to_polar_markers(
+        self,
         n_best: int = 1,
         marker_colors: str = ("red", "blue", "green", "orange", "purple"),
         lazy_output: bool = None,
@@ -1112,53 +1135,37 @@ class OrientationMap(DiffractionVectors2D):
             An list of markers for each of the n_best solutions
 
         """
-        (
-            r_templates,
-            theta_templates,
-            intensities_templates,
-        ) = self.simulation.polar_flatten_simulations(
-            signal_axes[1].axis, signal_axes[0].axis
-        )
-
         if lazy_output is None:
             lazy_output = self._lazy
+        if not lazy_output:
+            navigation_chunks = (5,) * self.axes_manager.navigation_dimension
+        else:
+            navigation_chunks = None
 
-        def marker_generator_factory(n_best_entry: int, r_axis, theta_axis):
-            theta_min, theta_max = theta_axis.min(), theta_axis.max()
-
-            def marker_generator(entry):
-                index, _, rotation, mirror = entry[n_best_entry]
-                index = index.astype(int)
-                mirror = mirror.astype(int)
-
-                r_ind = r_templates[index]
-                r = r_axis[r_ind]
-
-                theta_ind = theta_templates[index]
-                theta = theta_axis[::mirror][theta_ind] + np.deg2rad(rotation)
-
-                # Rotate as per https://github.com/pyxem/pyxem/issues/925
-                theta += np.pi
-
-                # handle wrap-around theta
-                theta -= theta_min
-                theta %= theta_max - theta_min
-                theta += theta_min
-
-                mask = r != 0
-                return np.vstack((theta[mask], r[mask])).T
-
-            return marker_generator
+        def vec2polar(vector):
+            # Bug with dtype of python float
+            vector = vector.astype(np.float64)
+            r = np.linalg.norm(vector[:, :2], axis=1)
+            theta = np.arctan2(
+                vector[:, 1],
+                vector[:, 0],
+            )
+            # flip y
+            theta = -theta
+            return np.vstack([theta, r]).T
 
         all_markers = []
         for n in range(n_best):
             color = marker_colors[n % len(marker_colors)]
-            markers_signal = self.map(
-                marker_generator_factory(n, signal_axes[1].axis, signal_axes[0].axis),
-                inplace=False,
-                ragged=True,
+
+            vecs = self.to_vectors(
+                n_best_index=n,
+                return_object=False,
                 lazy_output=True,
+                navigation_chunks=navigation_chunks,
             )
+            markers_signal = vecs.map(vec2polar, inplace=False, ragged=True)
+
             if "sizes" not in kwargs:
                 kwargs["sizes"] = 15
             markers = hs.plot.markers.Points.from_signal(